Added bottle, shoot bottle.

+++ src/main/scala/com/catinthedark/sszb/entity/AnimationWrapper.scala
@@ -2,4 +2,4 @@ package com.catinthedark.sszb.entity
 
 import com.badlogic.gdx.graphics.g2d.Animation
 
-class AnimationWrapper(val animation: Animation, val x: Int, val y: Int, var stateTime: Float = 0f )
+class AnimationWrapper(val animation: Animation, val x: Float, val y: Float, var stateTime: Float = 0f )
